fix(接口管理):修复接口定义调试另存为用例标题显示错误问题以及项目管理用户名称超长显示错乱问题
--bug=1038231 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001038231
--bug=1037880 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001037880
